refactor(project-structure): dividindo o projeto em controller, dto, modules e service. Movendo arquivos de testes para o diretorio test/
