+++ back-end/jest.config.ts
@@ -1,0 +1,14 @@
+import { Config } from 'jest';
+
+const config: Config = {
+  moduleFileExtensions: ['js', 'json', 'ts'],
+  rootDir: '.',
+  testMatch: ['**/test/**/*.spec.ts'], // 👈 olha os testes na pasta test/
+  transform: {
+    '^.+\\.(t|j)s$': 'ts-jest',
+  },
+  coverageDirectory: './coverage',
+  testEnvironment: 'node',
+};
+
+export default config;
