+++ back-end/src/modules/profile.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { ProfileController } from './profile.controller';
-import { ProfileService } from './profile.service';
+import { ProfileController } from '../controllers/profile.controller';
+import { ProfileService } from '../services/profile.service';
 import { PrismaService } from 'src/services/prisma.service';
 
 @Module({
